Fix deprecation (#430)
* Fix deprecation

I fixed the error disallowing different-type outcome containers. However, that introduced the deprecation warning again. This fixes the resulting deprecations warning.

* Increment path version

* remove assert

if probs don't sum exactly to 1, then we normalize in the last step anyways.

* fix binning imprecision and reintroduce assertion.

the assert statement did actually catch a mistake that was caused elsewhere, so it should stay in and keep doing its job.